+++ src/data_model/item.rs
@@ -1,15 +1,15 @@
-use super::vector::Vector;
+use crate::vector::Vector;
 use std::collections::HashMap;
 
 pub struct Item {
-    pub id: String,
-    pub vector: Vector,
-    pub metadata: Option<HashMap<String, String>>
+    id: String,
+    vector: Vector,
+    metadata: Option<HashMap<String, String>>
 }
 
 impl Item {
     pub fn new(id: String, vector: Vector, metadata: Option<HashMap<String, String>>) -> Self {
-        Item {id, vector, metadata}
+        Self { id, vector, metadata }
     }
 
     pub fn get_id(&self) -> &String {
