--- src/data_model/lib.rs
@@ -1,7 +1,0 @@
-pub mod vector;
-pub mod item;
-// pub mod store;
-
-pub use vector::Vector;
-pub use item::Item;
-// pub use store::VectorStore;
