+++ src/data_model/metrics/metric.rs
@@ -1,0 +1,5 @@
+use super::vector::Vector;
+
+pub trait SimilarityMetric {
+    fn compute(v1: &Vector, v2: &Vector) -> Option<f32>;
+}
